+++ day_02_i_was_told_there_would_be_no_math/src/puzzle/dimension.rs
@@ -1,5 +1,5 @@
 pub struct Dimensions {
-    pub(crate) length: i64,
-    pub(crate) width: i64,
-    pub(crate) height: i64,
+    pub(crate) length: usize,
+    pub(crate) width: usize,
+    pub(crate) height: usize,
 }
